FUSETOOLS2-184 - provide Diagnostic on Properties files
- diagnostic with precise range only for enum and invalid
unknownparameter (like for other files)
- no Quickfix supported, can be done in a separated PR, this one is huge
enough
