+++ src/test/resources/workspace/diagnostic/camel-valid.properties
@@ -1,0 +1,1 @@
+camel.component.timer.basicPropertyBinding=true
