+++ src/test/resources/workspace/diagnostic/camel-with-2-errors.properties
@@ -1,0 +1,2 @@
+camel.component.timer.unknownparameter=3
+camel.component.timer.anotherunknownparameter=3
