+++ src/test/resources/workspace/diagnostic/camel-with-boolean-error.properties
@@ -1,0 +1,1 @@
+camel.component.timer.basicPropertyBinding=invalidBoolean
