+++ src/test/resources/workspace/diagnostic/camel-with-integer-error.properties
@@ -1,0 +1,2 @@
+camel.component.timer.basicPropertyBinding=true
+camel.component.activemq.acknowledgementMode=InvalidInteger
