+++ src/test/resources/workspace/diagnostic/camel-with-invalid-enum.properties
@@ -1,0 +1,1 @@
+camel.component.activemq.errorHandlerLoggingLevel=WRN
