+++ src/test/resources/workspace/diagnostic/camel-with-unknownParameter.properties
@@ -1,0 +1,1 @@
+camel.component.timer.unknownparameter=3
